+++ src/Catalog/DeclinationSummary.php
@@ -75,12 +75,7 @@ final class DeclinationSummary
             return new DeclinationOption($data);
         }, $data['options']);
         $this->mainImage = isset($data['mainImage']) ? new Image($data['mainImage']) : null;
-        $this->company = new CompanySummary(
-            $data['company']['id'],
-            $data['company']['name'],
-            $data['company']['slug'],
-            isset($data['company']['image']) ? new Image($data['company']['image']) : null
-        );
+        $this->company = new CompanySummary($data['company']);
         $this->slug = $data['slug'];
         $this->categoryPath = array_map(static function (array $data): ProductCategory {
             return new ProductCategory($data);